Issue: SSW-5231 - [amx] Security cannot be modified
Problem: keys value in WiFi.Accesspoint.x.Security can't be modified when they are numerical [0-9]

Analysis: When setting a key having a numrical value in the data model, the value is read NULL in the writeHandler.
          In fact, the key_value type_id is considered by amx as AMXC_VAR_ID_INT32 instead of AMXC_VAR_ID_CSTRING.

Solution: use amxc_var_dyncast function instead of amxc_var_constcast to do the dynamic conversion of type.
GitOrigin-RevId: 27f7fd4f5d36e6eeca5e150803a55d85ca92219b