Issue: SSW-7360 - BSSID not properly set when external config source
Problem: BSSID is not properly updated in case when config comes
from external source.

Reason: there is no syncSSID SET called.

Solution: after a syncSSID GET is done in FSM, and init was not
done before, then do a syncSSID set to write back values.

Also updating control flow to move the syncApDm directly in
the "change event handler". This was previously done in delay
exec handler due to fact that initially function was called
as "parameter write handler", but is now always called in
"parameter change event handler".

GitOrigin-RevId: 6556e46860b9d133c041c730f1c439aea56e6dcd