Issue: SSW-7679 - Channel is not upgrade persistent
Problem:
- After the upgrade, radio channel is configured to its default value.

Cause:
- Prior to receiving the event triggered by pcm module upon setting
the radio channel, the syncData_Radio2OBJ function is invoked to initialize
the parameters of the created radio object.
- The restored value of the channel will be replaced by the value set through
the transaction initiated by the invocation of syncData_Radio2OBJ function.

Solution:
- During radio initialization, check whether the channel is already
configured. If it is, avoid to set the default channel value.
- Later upon receiving the write event, the internal context of the radio
will be updated, syncing its values with those on the DM object.
GitOrigin-RevId: aba613b28eaaf7a95446052eba7a9252e7d588ee