Issue: SSW-7781 - Wi-Fi is off after modifying default IPAddress
Problem: When using WebUI to change IPAddress, a called dora cycle is
executed to Disable/Enable all bridge ports. Since WiFi VAPs are ports
in the bridge they are following same behavior. The dora cycle is based
on the Status parameter change of the VAP, to re-enable it again.

Cause: Some times the Status do not go from Enabled to Disabled despite
that Enable parameter is set to false.

Fix: force to write value of the VAP status into DM without the need for
an intrenal old/new values difference.
GitOrigin-RevId: 70fbfce5d69a96792d967f7bdf26938cd4cb6452