Issue: SSW-7906 - fix station disabling HT/VHT/HE/BE if no WMM/QoS
Problem: when using a softmac (kernel mac80211) station to connect to
a VAP that do not support the NL80211_SET_QOS_MAP, wmm_enabled will be
set to 0 for that VAP, the sta will disable HT/VHT/HE.

Cause: the WMM IE is not present in the Beacon/Probe Response frames.

Fix: only set the hostapd wmm_enabled according to VAP WMMEnable parameter
when WMMCapability is detected and true, otherwise do not set
wmm_enabled parameter at all and let hostapd decide about it.
GitOrigin-RevId: c68d9a8b4c7470af30d384cc76822e00a9a1fa6f