Issue: SSW-8070 - MaxAssociatedDevices on radio is not working
Problem: The limit of stations per radio is not enforced.

Cause: Not implemented because hostapd does not provide a param
to limit the number of stations per radio. It only provides a
limit per bss.

Solution: Monitor station associations and dynamically update
the limit per bss in hostapd to honor both the configured limit
per bss and per radio.

Note that the solution is not immune to a race condition where
two station associate at the same time with different access points
on the same radio.

GitOrigin-RevId: 44c1eacfbd93e2e02f71b8b79377eb875dc21fe9