Issue: SSW-8377 - neighbour discovery value not set properly
Problem: neighbour discovery radio value not properly set at boot

Cause: when configuring the setting in FSM, the setting is read
while the radio & AP are down. This causes the neighbour discovery
calculation function to think there is no AP "up", and provide
wrong configuration.

Solution: instead of checking the "current state" of the APs,
check the desired state instead.
GitOrigin-RevId: 5c07e563f3da2d8f27822dc41d160c493b1afdb4