Issue: SSW-8601 - Adjust the fronthaul channel configuration upon backhaul connection
Problem: After connecting an endpoint, the active Virtual Access Point
(VAP) remains disabled, attempting to switch to a different channel
than the one used by the endpoint. In addition, the setChanspec RPC
performs actions even when an endpoint is connected.

Cause: Upon receiving the station connected event, only the current
chanspec is updated and the channel in the hostapd's configuration file
does not match the current radio channel.

Fix: Update the target chanspec when the current chanspec changes due to
an endpoint connection and ensure the radio has no endpoint connected
before continuing with the setChanspec RPC.
GitOrigin-RevId: 290aea2809c5132740e1b96d05de19509af00b75