Issue: SSW-8661 - increase 4way handshake retries.
Problem: The KickRoamingStation configuration does not have sufficient
time using 4-way handshake default retries, preventing the station
from disconnecting from the old vap.

Cause: The time allocated to receive the associated management frame
and detect a station's double association may be inadequate, which
contributes to the delay in disconnecting the station.

Fix: Extend the time allocated for the 4-way handshake when the
KickRoamingStation feature is enabled.
GitOrigin-RevId: 79cfa52fe5f8c4b639326ea0877ae11572a0580b