Issue: SSW-8674 fix auto generated mac overlap
Problem:
- reproduced when having rad6g, in 2nd position in radio list
- vap5gpriv and vap6gguest have same MAC

Cause:
- using statically configured  baseMacOffset does not consider the case
where the rad6g (requiring contiguous phy MACs for all BSSs) is in the
middle of radio list.
- rad mac is derived from previous radio basemac, while in the test, the
order of radio instance is different than the radio dev detection order.

Fix:
- remove static baseMacOffset and let pwhm shift as much as needed
baseMacs considering prev radios nbssReq and maxHwBss
- take previous radio from the dm object list (isof the device detection list)
as the MACConfig used to calculate the rad BaseMac is loaded in the
dm instance order.
GitOrigin-RevId: 22cf04c0aa3ad3cf1143c6048715d86236bfed47