Issue: SSW-8217 - update endpoint profile status
Problem: When setting a profile's parameters which is not the reference
one, the status of the profile is not update. When seeting this profile
as the profile reference of the endpoing, SSID's endpoint will not be
update to the one configured in the profile and keep using default SSID
(PWHM_SSIDX)

Cause: profile status must be at least in available status to update the
endpoint SSID

Fix: update profile status when Enable parameter is modified even if the
profile is not selected as the profile reference.
GitOrigin-RevId: fe770bc461113fd4399059d7396f4ed2e1f21b19