Issue: SSW-8774 - MACFilterAddressList modification does not work
Problem: when there is already a MACFilterAddressList configured
and we set it to a different value, the result is an empty
MACFilterAddressList.

Cause: the addition of the new entry and deletion of the old entry
in the MACFiltering.Entry table are done in a non-atomic way with
multiple callbacks involved and an intermediate state where the
table is considered empty, because the newly added entry is not
marked as valid yet. This intermediate state triggers resetting
the MACFilterAddressList to an empty string.

Solution: mark the newly added entry as valid as soon as the transaction
is successful, without waiting for the object creation handler to do it.

Note that the associated unit test which is added as part of this
commit does not catch this particular problem because the order of
callbacks is different between unit tests and target.

On target the sequence leading to the issue from the intermediate state is:
1. commit MACFilterAddressList=""
2. delayExec of s_updateMACFilterAddressList
3. s_updateMACFilterAddressList is called after ~10ms
4. the commit handler for MACFilterAddressList="" is called after ~20ms

In the unit test environment (where time is mocked by the test-toolbox):
1. commit MACFilterAddressList=""
2. delayExec of s_updateMACFilterAddressList
3. the commit handler for MACFilterAddressList="" is called immediately (no time travel needed)
4. s_updateMACFilterAddressList is called after 1ms of time travel because
   swla_delayExec_add actually starts a 1ms timer

This results in the MACFilterAddressList being empty on target but not
in unit tests.

With the fix we do not go through this inconsistent intermediate state
anymore, thus the problem does not occur on target anymore.

GitOrigin-RevId: 73380648652a8cc6d9063bb0e5037f79f77d44df