PPM-2923 Fix for no rekey packets seens when wpa_group_rekey intervals id changed from default
ISSUE: Rekey packets not triggered at mentioned interval

CAUSE: wpa_group_rekey interval value is updated in config but timer not restarted with the updated config as hostapd is not restarting timer on reconfig.

FIX: pwhm sends TOGGLE instead of SIGHUP to hostapd, so that hostapd updated the rekey interval as well as restarts the rekey timer, as a result rekey packets are triggered at specified intervals.

Closes: PPM-2923;

Signed-off-by: Mujibur Rahiman K <mujiburrahimank@maxlinear.com>
GitOrigin-RevId: 75adcc80bb1702fdec126ade8484acb825993039