Issue: SSW-8770 - WiFi bits not executed
Problem: FSM is not executing all bits on certain operations

Cause: When FSM is non idle, and a change is pushed, then
the auto commit manager will not trigger, not actually
starting the count down for a commit.

If this happens after the DEPENDENCY step, then the commit
flags are present, but never actually commit, as the Commit
Pending is never incremented.

Solution: Make the auto commit manager trigger as soon
as no commit pending is present.

Note that this cleans up a previous attempt to fix this, by
checking the CSC. However, the CSC are a set of bits that
are only updated at some very brief point in time, and
should not be used afterwards. Adding a check here may
cause FSM to trigger unintentially multiple times.

GitOrigin-RevId: 124dd7fea5d64012cb1cd11081c178788c20bcd2