Issue: SSW-8770 - WiFi FSM toggling of Vendor FSM
Problem: A vendor FSM is continuously in FSM commit loop

Cause: The vendor has a custom FSM, but a radio is detected that
is not handled by the vendor module.

This radio then causes the wifiGen_fsm to subscribe to vap changes.
The change handler however is called for all vap changes, for
all FSMs, and not just for the vaps that are handled by the
generic FSM.

Solution: check on callback handler that the callback is actually
from an AP on a radio that the generic FSM is handling.
Return otherwise.
GitOrigin-RevId: 9c36e6d4bcd224c447fd517591822430ddab5d40