Issue: SSW-8770 - Extra radio in data model
Problem: 3 radios seen when only 2 radios present

Cause: On boot, there are two interfaces on a single phy.
This causes the nl80211 add radio logic to add two radio's for
that phy, because it only expects a single interface per logical phy.

The logic must be able to handle multiple radio's are phy, as
standard MLO design works by having a single nl80211 phy, but multiple
interfaces

Solution: check if the system is a single phy, or multi phy setup,
and if multi-phy, only add a single interface per phy
GitOrigin-RevId: 76b9aa51f8682dd66968c8b0079f1a5fcfec2d9a