Issue: SSW-8756 - Operating standard not updated after device reassoc
Problem: When the radio operating standard is changed from "bgn" to "ax"
the associated device is disassociated and reassociated. After reassociation
the operating standard of the device is "ax" but the value is not updated in
the data model.

Cause: There is a check to determine whether the operating standard was set
by the driver or if it needs to be set from the association request IEs. But
this test only works once and does not work after a reassociation.

Solution: Introduce a separate flag to remember whether the operating standard
and bandwidth were set by the driver or from the assoc request IEs.
GitOrigin-RevId: c958acb210703733c9ffa5ac3cbc39e2c86b4074