[pwhm] Prevent overwrite of associated device data with empty values
When copyAssocDevInfoFromIEs is called from a probe request, not all the
IEs are present and most of the data is empty. Because of this, we overwrite
data that was parsed from the association request IEs. On the CLI we can see
that the following are zeroed after the device sends a probe request:
MaxDownlinkRateSupported, MaxRxSpatialStreamsSupported,
MaxTxSpatialStreamsSupported, MaxUplinkRateSupported.

GitOrigin-RevId: 1d49665fd7fe690b7a706e6061355d4f41a9eadc