Issue: SSW-8927 - fix SSID sync looping
Problem: synchronisation between Endpoint/AccessPoint and SSID is
looping.

Cause: when modifying the AccessPoint or Endpoint Enable parameter,
a synchronization with the corresponding SSID is triggered, which in
turn attempts to synchronize back the SSID Enable parameter with
AccessPoint/Endpoint.

Fix: Synchronize the internal SSID enable value directly with the new
Endpoint/Access Point's enable value, ensuring that when the writer
handler is invoked, no action is taken.
GitOrigin-RevId: ca3f1e58883671e6c45674d660c1e526865b7252