PPW-34 Fix 6G VAP advertised in neighbor radios RNR when ssid is hidden
Issue: When 6G VAP is conifgure with hidden ssid and AutoAddNeigh enabled, 2.4G and 5G radio still advertise it in RNR IE.

Root Cause: 6G VAP neigh instance is not added/deleted when its SSID advertisement mode is changed.
And adding neigh instance is not blocked when SSID is hidden.

Fix: Refresh 6G VAP neigh when hidden ssid mode changed and block creating or updating 6G neigh instance when ssid is hidden.
This mandatory because it will also prevent it from being created when interfaces are restarted/toggled.

Signed-off-by: Nikita Chernikov <nchernikov@maxlinear.com>