PPW-59 Returning from the s_doUpdateHostapd without doing the Reload
Issue: When Radio is disabled in single hostapd instance, Reload hostapd is happening.
Root cause: The check for the hostapd process running is not modified for the single hostapd instance.
Proposed fix: Determine the hostapd is running in the Group of hostapd by checking whether that secDaemon is enabled or not.
Signed-off-by: Vaish,Prabhav Kumar pvaish@maxlinear.com
Closes PPW-59
GitOrigin-RevId: 4869c2a15fcc1b6edbc6a97544b56f5a6c99aa28