Issue: PPW-72 mxl-allowing-vap-addition-more-than-permissible-limit
Problem: Allowing vap addition more than the permissible limit or hardware limit leading to radio down.
Example: In 6G if we add more than 32 Vaps if the hardware limit is 32,
interface is going down.

Root Cause: There is no checks for number of vaps are added.

Fix: check for the max hardware limit of vaps and current vaps, if limit is exceeded return.

Signed-off-by: Rahul R <rahulr@maxlinear.com>
GitOrigin-RevId: 71a23d2b7f75bd32c09ac07ada72c3bbd9204310