SSID Change on Main VAP causes deauthentication on both Main and Guest VAPs
Issue:
After boot up or prplmesh_whm restart, changing the SSID
of the main VAP causes disconnection of stations connected
to both the main and guest VAPs. However, changing the SSID
of the main VAP a second time only disconnects the station
connected to the main VAP, which is the expected behavior.

Root Cause:
    - During init, PWHM updates the config id only for the
master vaps.
    - A Mismatch occurs between the old and the new config
id of each bss.
    - Hostapd config reloads for both Main and Guest Vaps.
    - Therefore, deauth is sent to the stations associated
with both the main and guest vaps.

Fix:
The config Id is updated for all bss during init.

Closes PPW-98

Signed-off-by: svenkatachalam <svenkatachalam@maxlinear.com>