Issue: PPM-3034 PWHM Wi-Fi hostap updated security configuration for Wi-Fi 7
Wi-Fi 7 introduces new mandatory security methods and features
- AKM-24
- GCMP-256
- Beacon Protection

In the hostapd / wpa_supplicant configuration, if the AP is linked to a
Wi-Fi 7 operating standard radio and has a WPA3 or WPA3-Transition security
mode enabled, then:
- Append SAE-EXT-KEY to the wpa_key_mgmt (resp. key_mgmt in supplicant)
- Append GCMP-256 to the wpa_pairwise (resp. pairwise in supplicant)
- Add beacon_prot=1 in the hostapd configuration.

Assumptions:
- Any Wi-Fi 7 capable radio implements AKM-24, GCMP-256 and beacon
protection as part of the mandatory features of Wi-Fi 7 specification.
- AKM-24, GCMP-256 only is not expected to be configured by an end-user.
So a AKM-8/AKM-24 & CCMP/GCMP-256 transition mode is currently always
expected for now.

Signed-off-by: NGUYEN DINH HIEN Michael Thien Bao <michael.nguyen@sagemcom.com>

Closes PPM-3034