Issue: HGWPR-32104 Received error during boot
Problem: Significant number of error traces during boot: transaction fails because transaction is requested to create an object that already exists.

Cause: Due to a high amount of addNeigh requests, there seems to be a desync between the execution of the create object transaction, and the post transaction internal object creation handler. When then two addNeigh on same object are called very closely together, the first creates the transaction, executes transaction and creates object. The second call is then executed before internal object is created, which means that, when looking for the internal neigh, no object is found, while it actually exists.

Solution: instead of looking for internal model, just check the bus data model to see if object exists.
GitOrigin-RevId: 0fbfa6e102f733870479db17aac8c5303e0c7dfe