Issue: SSW-9557 - pWHM with external configation slow to start
Problem: pWHM boots wifi slower than expected

Partial cause: during initial bringup of pWHM on target, it was
seen that configuration was sometimes very slow to deploy.

Reasons unclear, but an initDone check was added, but no
clear check whether any init was actually done.

Current state is that config deploy seems to have significantly
sped up, so issue is no longer relevant and now just a slowdown

Solution: add a check whether SSID is default, and only
consider that vaps are not yet initialized, if SSID is still
in default value.

Also clarify check function on why bus has to be differnet from
NULL, as it indicates a dummy VAP.

GitOrigin-RevId: 03cc4ec66cbe1f7e6e6deed7c0e0e9e9c443b4d2