Issue: SSW-9680 - EP fails to connect after WPS
Problem: After doing WPS with a WPA2-WPA3 AP, EP fails to reconnect

Cause: on certain board, WPS returns a 64hex psk. When this
key is used as profile key, and configure the security as
WPA2-WPA3, then wpa_supplicant fails to connect

Solution: When WPS returns a 64hex psk, do not add WPA3 mode
to security returned, as the 64 hex psk can not be used for
sae.

If 64char psk is provided, and it is not hex, send out WPS failure.

GitOrigin-RevId: bc2c37dc628f446f9434f5e6e0afd0c7bfd62c25