Issue: PPW-98 SSID Change on Main VAP Causes Deauthentication on Both Main and Guest VAPs
Problem: Changing the SSID of one the VAP on same radio which has multiple VAP's disconnect the clients on the other VAP's. For example Client are connected to Guest VAP (2.4GHz), and SSID is changed on the Main VAP (2.4Ghz) this will lead to disconnect of clients on Guest VAP(2.4Ghz)

Cause: Due to mismatch in the config_id of the VAP between hostapd config (/tmp/wlanX_xxx.conf) and updated to hostapd over wpa control interface. Mismatch already happens when we enable the Guest VAP, the config_id is updated (newly generated) and updated in the hostapd config (/tmp/wlanX_xxxx.conf) but not passed on to hostapd over wpa control interface. Later when the Main VAP SSID is changed which leads to reload of config hostapd_reload_config()@hostapd and results in hostapd_clear_old_bss()@hostapd on the guest VAP. Since the hostapd config (/tmp/wlanX_XXXX.conf) has the config_id updated but the NOT hostapd data structures.

Fix: Update the config_id for hostapd over the wpa ctrl interface in addition to update in the hostapd config.

Tests:
With clients already connected to guest, Change Priv SSID and ensure no impact on clients connected to guest :OK
With clients already connected to guest , Then Also connect clients to Priv VAP and then change the SSID of GUEST VAP No disconnect on Priv VAP :OK
With Clients connected to guest and priv VAP, enable the disable the VAP's and see clients on other VAP are not disconnectded :OK
Disconnect and Reconnect the clients from VAP and ensure other VAP clients are not impacted :OK
Connect clients to both VAP (priv/guest) change the keyPassPhrase of the both VAP one by one, other VAP clients not disconnected :OK
Disabled the RNR on priv vap, no disconnect is seen on VAP, rnr and config_id is sent over wpa ctrl interface and hostapd becaon update is seen for priv vap :OK
SET "WiFi.AccessPoint.2.MBOEnable=1" on VAP, ensure the no disconnect is seen on any VAP and the correct vaule is read by hostapd on config reload :OK
Monitor the clients disconnection by disabling auto-connect on client and iw wlanX.Y station dump

Signed-off-by: Ashutosh Shandilya <ashutosh.shandilya_ext@softathome.com>
GitOrigin-RevId: 98794666a1239b7bba4c3c49b4317109d737efd1