Issue: SSW-9574 - remove reordering list of bss.
Problem: in the wld_hostapd_createConfig() a reorder of the active vap
is done before creating the hostapd configuration. no need to do so
since vaps are already ordered according to there creation in the
Datamodel, and BSSID are following this order.

Fix: remove the reorder of vap list, and just use the order done alredy
by s_initConfig()
GitOrigin-RevId: 1167481b2ee7b17c67fae2e210860ee1364d59b5