PPW-306: Endpoint unable to connect when OWE mode is configured
RCA: pwhm is giving configurations to the wpa_supplicant
in the wpa_supplicant config file. For the OWE mode,
it is not implemented which in turn made it failure.

Fix: Handle the OWE security mode such that when OWE is
configured, the right configurations are given to the
wpa_supplicant

Tests Performed: Post fix able to see that the endpoint
was able to connect with the gateway's backhaul vap.

PPW-306

Signed-off-by: bsrinivasan <bsrinivasan@maxlinear.com>
GitOrigin-RevId: 3596cf4740096342709be23beac72136782fcb91