Issue: SSW-9901 - fix crash with invalid obj ref.
Problem: pwhm is crashing when trying to access an object reference
stored in a associated device object.

Cause: pointer is not valid (0x8f)! The object datamodel reference of
the associated device seems to be valid, the associated device mapped
object seems valid too, but the reference stored within the mapped
associated device internal object is invalid.

Fix: try to check the pAD->object is equal the object reference in the
datamodel to validate operation. This is already done by wld_ad_fromObj()
method, so use that instead of dereferencing the priv pointer directly.
GitOrigin-RevId: ef0b83bf6893a8d43d8d7826dec9439bb60ce167