Issue: SSW-9997 - Auto bandwidth selection does not work on channel 6/213
Problem: In the 6GHz band, when setting
  AutoChannelEnable=0
  Channel=213
  OperatingChannelBandwidth=Auto
  AutoBandwidthSelectMode=MaxAvailable
pwhm selects bandwidth 160 MHz while it should use 320 MHz.

Cause: When 320-1 MHz is not available, pwhm takes 160 MHz without checking for
320-2 Mhz.

Solution: Check both 320-1 and 320-2 MHz before resorting to 160 MHz. This means
that wld_chanmgt_getAutoBwExt does not only select the bandwidth but also the
extension (low or high). Thus it returns a swl_radBw_e instead of a
swl_bandwidth_e.
GitOrigin-RevId: 0989688150112dd9205aace67a7b0ac83c9f8c17