Issue: SSW-9939 - stop a crashed daemon
Problem: unable to stop a crashed daemon.

Cause: pre-stop checks are always checking whether a daemon is running,
not whether it's enabled. Since a crashed daemon is by definition
not running but IS enabled, it is impossible to actually
turn it off.

Solution: move checks to prevent stop call on enable instead of
running.
GitOrigin-RevId: 602a96f4fb82430d9ac297c2d489b09020a04b43