Issue: SSW-10117 - WPS Pin update in infite loop.
Problem: FSM commit triggering in loop.

Cause: Self-Pin is being updated in a loop on repeater. The
cause seems to be, that one of the VAPs is syncing back the
SelfPIN, while at the same time, higher layer network sync
software is updating the repeater with HGW pin.
The write back seems a bit slow, which then allows for
an infite loop to get triggered.

An alternate trigger is to just generate self pin
a few times.

Solution:
* WPS SelfPin is no longer written back on syncBack
* To avoid looping by generating self pin, it shall
be impossible to write back the previous PIN immediately.
This then immediately breaks the loop, avoiding infite loop
scenario's.
GitOrigin-RevId: 9bc186d02d984b13ad25c6c2c883ffc1049a3bb3