Issue: SSW-10149 - Radio up while not able to perform traffic.
Problem: radio is up while not able to do network traffic.
This violates the TR-181 definition of radio up.

Cause: radio is set to status up as soon as it is able to
send management traffic, i.e. while an endpoint is connecting
for example. However, at this time, the radio is only able
to send management traffic, but no network traffic.

Solution: in order to have radio considered up, have at least
a single VAP enabled, or an endpoint associated, so
radio can actually do network traffic.
GitOrigin-RevId: add13d41b30df10134321b376ef15e829409de41