PPW-404: Update backhaul credentials for EM test
In the EM R5 certification test, we are required to
onboard an external agent to an external controller
through us (DUT).

The backhaul configs are not populated in the
respective position leading to failure in onboarding
an external wireless agent.

Add changes to supply the backhaul credentials.

UT: Ran MAP-4.2.3_BHWIFI and it is passing.

PPW-404

Signed-off-by: bsrinivasan <bsrinivasan@maxlinear.com>