Issue: SSW-10179 - Endpoint connection is failing
Problem:
The 4 way handshake fails because the endpoint is unable to transmit frames in 4MAC mode.

Cause:
wpa_supplicant is unable to dynamically enbale 4addr on managed interfaces.
When the managed interface is a mld interface, enabling the 4addr after
association is not yet properly handled by the kernel mac80211.

Fix:
Enable the 4mac mode on the managed interface if the MultiAP is enabled.
GitOrigin-RevId: 17eccb1d48cafbf5495b986894d6bd57a62e96c8