Issue: SSW-10264 - crash in pwhm
Problem: Double free in s_scanResultsCb due to double trigger of this callback.

Cause: In s_scanResultsCb, multiple events are triggered in pwhm and some of
them propagate synchronously (without going back to the event loop). This leads
to sending new async nl80211 requests while the s_scanResultsCb has not returned
yet. One of these async requests blocks for 10 seconds and times out, which
triggers the cleanup of expired requests in wld_nl80211, and that includes the
scan request which has not been cleaned up yet. The s_scanResultsCb is thus
called a second time for the same request, which leads to double free.

Solution: Add a flag in the nl80211 request context to indicate whether the
callback is currently running on the stack. If yes, avoid cleaning up this
request. The request will be cleaned up when the callback returns, or upon the
next socket read after that.
GitOrigin-RevId: 79c95af2604cf125e969e870813285b723cd1d1d

(cherry picked from commit 8aef3b0c9d6f7a99ded47e7efbb0c8fe895ba4f6)
Signed-off-by: Frederik Van Bogaert <frederik.vanbogaert@mind.be>