Issue: PPM-3041 WiFi Sensing phase 2a step2.1 - support vendor data types
Problem:
We should use Maxlinear specific OUI in the supported data type list
(e.g. AC:9A:96)

Maxlinear recommends to use the value AC9A9680. The last byte is set
to 80 and not 01, since this is not a standard 11bf CSI data, but
vendor specific.

Cause:
PWHM accepts only a list aof predefined values.

Solution:
Remove implicit validator of SupportedSensingDataTypes , as the values
are set as capabilities by pwhm/vdrMod.
The default values can be set in the feed level.

Closes: PPM-3041

Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>