Issue: SSW-10465 PPW 491 : Fix SSID update on 6GHz band
Problem:
in case of MBSSID, the UPDATE_BEACON command is sent
to the 'beaconing interface'

for certain targets, pwhm selects the dummy interface.
as a result the UPDATE_BEACON has no effect : the SSID in
the air does not change

Problem was observed and reported on 6GHz band in PPW-491 because
6GHz band does not enable WPS.
It is not a 6GHz problem, for the following reason:
When WPS is enabled, the hostapd refresh is 'deeper',
so the problem is typically not seen on 2G4 and 5G.
When disabling WPS on any of these bands, the behavior is
the same as on 6GHz : SSID is not updated in the air.

Solution:
fix interface selection for UPDATE_BEACON hostapd command:
avoid sending to dummy vap on these targets

Issue:SSW-10465
Issue:PPW-491

Signed-off-by: Iacob Juc <iacob.juc_ext@softathome.com>
GitOrigin-RevId: b8e0e55b607b430a83070afb30f6a01f2c00702a