pWHM: update secure daemon group member state
Problem: When disabling a member of a secure daemon group, when this
group is waiting for all member to became startble, the retry callback
checking status of memebers do not update status.
This is leading to no secure daemon started, and infinit retries to
start it waiting for a disabled member to be startable.

Cause: Callback is not updating members status by it self.

Fix: Update members status to current status before checking.
GitOrigin-RevId: 4d8edabb6511498db09ea76895cd547f457f4ce6
(cherry picked from commit 28b8391ee58779d7855123bdb86c1ad99d1a5bd0)
Signed-off-by: Frederik Van Bogaert <frederik.vanbogaert@mind.be>