Issue: SSW-10902 - check the BSSID validity
Problem: If the odl contains an AccessPoint referencing a wrong SSID
with null bssid, hostapd keeps restarting in loop due to an invalid
configuration.

Cause: null BSSID for a BSS is not valid for hostapd.

Fix: check the validity of the AccessPoint's BSSID before adding a new
BSS section onto hostapd configuration file.
GitOrigin-RevId: f4fadcaa19a5803355cc8bd63195ce87bfd78527