Issue: SSW-10964 - Radio invalid status
Problem:
The radio remains in a down state, despite one of its VAPs being active.

Cause:
hapd main interface setup completed is not currently processed as the
AP-ENABLED event is not received on the first wpactrl interface.

Fix:
Handle the AP-ENABLED event when received through the first ready wpactrl
interface.
GitOrigin-RevId: 2ec325ef6be83356fce84e88a49b63f2b2be5bd0