Issue: SSW-11018 - Limit 6 GHz scan to preferred scanning channels
On repeaters with a 6 GHz uplink, the time to scan the entire 6 GHz band to find
the gateway can be long. Here we introduce a configurable parameter to limit the
scan to preferred channels, as defined in 80211ax-2021 section 26.17.2.3. The
parameter is false by default as safer option. In deployments where the backhaul
links are known to use PSCs, the option should be enabled to allow faster
establishment of the 6 GHz uplink.

GitOrigin-RevId: d2f0b69295c7aed792a4812c421f77028a55b513