Issue: SSW-11165 - Avoid infinite recursion in wifiGen_refreshEpConnStatus
Problem: When an EndPoint connection event is received from wpa_supplicant at
the same moment the EndPoint gets disabled in the datamodel (e.g. because
switching backhaul), we enter an infinite recursion in
wifiGen_refreshEpConnStatus which leads to a stack overrun.

Cause: wld_endpoint_sync_connection overwrites the connection status to
EPCS_DISABLED when the EndPoint is configured disabled. This causes a mismatch
between the configured value and wpa supplicant status. In this case
wifiGen_refreshEpConnStatus regenerates an EP connected event, which turns into
an infinite recursion.

Solution: Do not call wld_endpoint_sync_connection when handling events. This
function is meant for configuration handling. In case of events, call
wld_endpoint_setConnectionStatus directly and avoid overwriting the connection
status based on configuration. The configuration will be propagated afterwards
by the FSM.
GitOrigin-RevId: 1f2c6391dd0a934279a74ded910834d2c7f5fa7d