wld_rad_nl80211: set other_bss_time when timeRxInBss is equal to or greater than 0
Currently, the `other_bss_time` in `wld_airStat_t` is calculated only
when `timeRxInBss` > 0. Thus, when there is no change in `timeRxInBss`
between RadioAirStats fetches by upper layers, this skips calculating
it and returns `other_bss_time` as 0, which is incorrect.

Thus, remove the `timeRxInBss` check, as it cannot be negative being
an unsigned integer.

PPW-719

Signed-off-by: Sagar Reddy <snandanuri@maxlinear.com>