Issue: SSW-11181 - Avoid connecting EP to a bss on the wrong band
Problem:
wpa_supplicant may attempt to connect to a BSS on a frequency not configured
in the configuration file

Cause:
The supplicant initiates a scan over a predefined set of frequencies and proceeds
to connect to the BSSID corresponding to the target SSID. Due to cached scan results
on the driver side, BSSIDs on frequencies outside the supplicant's configured list
may still be returned and considered.

Fix:
Set ignore_old_scan_res in the config file to prevent the supplicant from using
outdated scan results.
GitOrigin-RevId: d3d10d95f258c5bd905a8025a6acdb808f56cd6a